+++ src/utils.cpp
@@ -63,8 +63,8 @@ std::vector<std::vector<uint64_t>> utils::gsw_gadget(size_t l, uint64_t base_log
   // Create RGSW gadget.
   std::vector<std::vector<uint64_t>> gadget(rns_mod_cnt, std::vector<uint64_t>(l));
   for (size_t i = 0; i < rns_mod_cnt; i++) {
-    const uint128_t mod = coeff_modulus[i].value();
-    uint128_t pow = 1;
+    const uint64_t mod = coeff_modulus[i].value();
+    uint64_t pow = 1;
     for (int j = l - 1; j >= 0; j--) {
       gadget[i][j] = pow;
       pow = (pow << base_log2) % mod;
